+++ sdk/search/azure-search-documents/azure/search/documents/_internal/_search_index_document_batching_client.py
@@ -4,15 +4,15 @@
 # license information.
 # --------------------------------------------------------------------------
 from typing import cast, List, TYPE_CHECKING
+import time
 import threading
-from typing_extensions import Protocol
 
 from azure.core.tracing.decorator import distributed_trace
-from azure.core.exceptions import HttpResponseError
-from .._api_versions import validate_api_version
+from azure.core.exceptions import HttpResponseError, ServiceResponseTimeoutError
 from ._utils import is_retryable_status_code
+from ._search_index_document_batching_client_base import SearchIndexDocumentBatchingClientBase
 from ._generated import SearchIndexClient
-from ._generated_serviceclient import SearchServiceClient
+from ..indexes import SearchIndexClient as SearchServiceClient
 from ._generated.models import IndexBatch, IndexingResult
 from ._search_documents_error import RequestEntityTooLargeError
 from ._index_documents_batch import IndexDocumentsBatch
@@ -24,24 +24,7 @@ if TYPE_CHECKING:
     from typing import Any, Union
     from azure.core.credentials import AzureKeyCredential
 
-class PersistenceBase(Protocol):
-    def add_queued_actions(self, actions, **kwargs):
-        # type: (*str, List[IndexAction], dict) -> None
-        pass
-
-    def add_succeeded_action(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
-
-    def add_failed_action(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
-
-    def remove_queued_action(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
-
-class SearchIndexDocumentBatchingClient(HeadersMixin):
+class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, HeadersMixin):
     """A client to do index document batching.
 
     :param endpoint: The URL endpoint of an Azure search service
@@ -50,49 +33,36 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
     :type index_name: str
     :param credential: A credential to authorize search client requests
     :type credential: ~azure.core.credentials.AzureKeyCredential
+    :keyword bool auto_flush: if the auto flush mode is on. Default to True.
     :keyword int window: how many seconds if there is no changes that triggers auto flush.
-        if window is less or equal than 0, it will disable auto flush
-    :keyword int batch_size: batch size. Default to 1000. It only takes affect when auto_flush is on
-    :keyword persistence: persistence hook. If it is set, the batch client will dump actions queue when it changes
-    :paramtype persistence: PersistenceBase
+        Default to 60 seconds
+    :keyword hook: hook. If it is set, the client will call corresponding methods when status changes
+    :paramtype hook: IndexingHook
     :keyword str api_version: The Search API version to use for requests.
     """
     # pylint: disable=too-many-instance-attributes
-    _ODATA_ACCEPT = "application/json;odata.metadata=none"  # type: str
-    _DEFAULT_WINDOW = 0
-    _DEFAULT_BATCH_SIZE = 1000
 
     def __init__(self, endpoint, index_name, credential, **kwargs):
         # type: (str, str, AzureKeyCredential, **Any) -> None
-
-        api_version = kwargs.pop('api_version', None)
-        validate_api_version(api_version)
-        self._batch_size = kwargs.pop('batch_size', self._DEFAULT_BATCH_SIZE)
-        self._window = kwargs.pop('window', self._DEFAULT_WINDOW)
-        self._auto_flush = self._window > 0
+        super(SearchIndexDocumentBatchingClient, self).__init__(endpoint, index_name, credential, **kwargs)
         self._index_documents_batch = IndexDocumentsBatch()
-        self._endpoint = endpoint  # type: str
-        self._index_name = index_name  # type: str
-        self._index_key = None
-        self._credential = credential  # type: AzureKeyCredential
         self._client = SearchIndexClient(
             endpoint=endpoint, index_name=index_name, sdk_moniker=SDK_MONIKER, **kwargs
         )  # type: SearchIndexClient
         self._reset_timer()
-        self._persistence = kwargs.pop('persistence', None)
 
-    def _cleanup(self, flush=True, raise_error=False):
+    def _cleanup(self, flush=True):
         # type: () -> None
         """Clean up the client.
 
         :param bool flush: flush the actions queue before shutdown the client
             Default to True.
-        :param bool raise_error: raise error if there are failures during flushing
-            Default to False which re-queue the failed tasks and retry on next flush.
-        :raises: ~azure.core.exceptions.HttpResponseError
+        :raises: ~azure.core.exceptions.HttpResponseError,
+                 ~azure.core.exceptions.ServiceResponseTimeoutError
+                 ~azure.search.documents.RequestEntityTooLargeError
         """
         if flush:
-            self.flush(raise_error=raise_error)
+            self.flush()
         if self._auto_flush:
             self._timer.cancel()
 
@@ -111,54 +81,50 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         """
         return self._index_documents_batch.actions
 
-    @property
-    def succeeded_actions(self):
-        # type: () -> List[IndexAction]
-        """The list of currently succeeded index actions in queue.
-
-        :rtype: List[IndexAction]
-        """
-        return self._index_documents_batch.succeeded_actions
-
-    @property
-    def failed_actions(self):
-        # type: () -> List[IndexAction]
-        """The list of currently failed index actions in queue.
-
-        :rtype: List[IndexAction]
-        """
-        return self._index_documents_batch.failed_actions
-
-    @property
-    def batch_size(self):
-        # type: () -> int
-        return self._batch_size
-
     def close(self):
         # type: () -> None
-        """Close the :class:`~azure.search.SearchClient` session.
+        """Close the :class:`~azure.search.documents.SearchClient` session.
+
+        :raises: ~azure.core.exceptions.HttpResponseError,
+                 ~azure.core.exceptions.ServiceResponseTimeoutError
+                 ~azure.search.documents.RequestEntityTooLargeError
 
         """
         self._cleanup(flush=True)
         return self._client.close()
 
-    def flush(self, raise_error=False):
-        # type: (bool) -> None
+    @distributed_trace
+    def flush(self, timeout=86400):
+        # type: (int) -> bool
         """Flush the batch.
 
-        :param bool raise_error: raise error if there are failures during flushing
-            Default to False which re-queue the failed tasks and retry on next flush.
-        :raises: ~azure.core.exceptions.HttpResponseError
+        :param int timeout: time out setting. Default is 86400s (one day)
+        :return: True if there are errors. Else False
+        :rtype: bool
+        :raises: ~azure.core.exceptions.HttpResponseError,
+                 ~azure.core.exceptions.ServiceResponseTimeoutError
+                 ~azure.search.documents.RequestEntityTooLargeError
         """
-        # get actions
+        has_error = False
+        begin_time = int(time.time())
+        while len(self._index_documents_batch.actions) > 0:
+            now = int(time.time())
+            remaining = timeout - (now - begin_time)
+            if remaining < 0:
+                raise ServiceResponseTimeoutError("Service response time out")
+            result = self._process(timeout=remaining)
+            if(result):
+                has_error = True
+        return has_error
+
+    def _process(self, timeout=86400):
+        # type: (int) -> bool
         actions = self._index_documents_batch.dequeue_actions()
         try:
-            results = self._index_documents_actions(actions=actions)
-            # re-queue 207:
+            results = self._index_documents_actions(actions=actions, timeout=timeout)
             if not self._index_key:
-                client = SearchServiceClient(self._endpoint)
-                kwargs = {"headers": self._merge_client_headers({})}
-                result = client.indexes.get(self._index_name, **kwargs)
+                client = SearchServiceClient(self._endpoint, self._credential)
+                result = client.get_index(self._index_name)
                 if not result:
                     # Cannot find the index
                     self._index_key = ""
@@ -170,33 +136,27 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
 
             has_error = False
 
+            # re-queue 207:
             for result in results:
-                action = [x for x in actions if x.get(self._index_key) == result.key]
-                if is_retryable_status_code(result.status_code):
-                    self._index_documents_batch.enqueue_actions(action)
-                    has_error = True
-                elif result.status_code in [200, 201]:
-                    if self._persistence:
-                        self._persistence.remove_queued_action(action)
-                        self._persistence.add_succeeded_action(action)
-                    self._index_documents_batch.enqueue_succeeded_actions(action)
-                else:
-                    if self._persistence:
-                        self._persistence.remove_queued_action(action)
-                        self._persistence.add_failed_action(action)
-                    self._index_documents_batch.enqueue_failed_actions(action)
-                    has_error = True
-
-            if has_error and raise_error:
-                raise HttpResponseError(message="Some actions failed. Failed actions are re-queued.")
-
-        except Exception:   # pylint: disable=broad-except
-            # Do we want to re-queue these failures?
-            self._index_documents_batch.enqueue_actions(actions)
-            if raise_error:
+                try:
+                    action = next(x for x in actions if x.additional_properties.get(self._index_key) == result.key)
+                    if result.succeeded:
+                        self._succeed_callback(action)
+                    elif is_retryable_status_code(result.status_code):
+                        self._retry_action(action)
+                        has_error = True
+                    else:
+                        self._fail_callback(action)
+                        has_error = True
+                except StopIteration:
+                    pass
+            return has_error
+        except Exception:  # pylint: disable=broad-except
+            for action in actions:
+                self._retry_action(action)
                 raise
 
-    def _flush_if_needed(self):
+    def _process_if_needed(self):
         # type: () -> bool
         """ Every time when a new action is queued, this method
             will be triggered. It checks the actions already queued and flushes them if:
@@ -212,7 +172,7 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         if len(self._index_documents_batch.actions) < self._batch_size:
             return
 
-        self.flush(raise_error=False)
+        self._process()
 
     def _reset_timer(self):
         # pylint: disable=access-member-before-definition
@@ -232,9 +192,8 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_upload_actions(documents)
-        if self._persistence:
-            self._persistence.add_queued_actions(actions)
-        self._flush_if_needed()
+        self._new_callback(actions)
+        self._process_if_needed()
 
     def add_delete_actions(self, documents):
         # type: (List[dict]) -> None
@@ -244,9 +203,8 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_delete_actions(documents)
-        if self._persistence:
-            self._persistence.add_queued_actions(actions)
-        self._flush_if_needed()
+        self._new_callback(actions)
+        self._process_if_needed()
 
     def add_merge_actions(self, documents):
         # type: (List[dict]) -> None
@@ -256,9 +214,8 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_merge_actions(documents)
-        if self._persistence:
-            self._persistence.add_queued_actions(actions)
-        self._flush_if_needed()
+        self._new_callback(actions)
+        self._process_if_needed()
 
     def add_merge_or_upload_actions(self, documents):
         # type: (List[dict]) -> None
@@ -268,15 +225,15 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_merge_or_upload_actions(documents)
-        if self._persistence:
-            self._persistence.add_queued_actions(actions)
-        self._flush_if_needed()
+        self._new_callback(actions)
+        self._process_if_needed()
 
-    @distributed_trace
     def _index_documents_actions(self, actions, **kwargs):
         # type: (List[IndexAction], **Any) -> List[IndexingResult]
         error_map = {413: RequestEntityTooLargeError}
 
+        timeout = kwargs.pop('timeout', 86400)
+        begin_time = int(time.time())
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         try:
             index_documents = IndexBatch(actions=actions)
@@ -286,21 +243,31 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
             if len(actions) == 1:
                 raise
             pos = round(len(actions) / 2)
+            now = int(time.time())
+            remaining = timeout - (now - begin_time)
+            if remaining < 0:
+                raise ServiceResponseTimeoutError("Service response time out")
             batch_response_first_half = self._index_documents_actions(
                 actions=actions[:pos],
                 error_map=error_map,
+                timeout=remaining,
                 **kwargs
             )
-            if batch_response_first_half:
+            if len(batch_response_first_half) > 0:
                 result_first_half = cast(List[IndexingResult], batch_response_first_half.results)
             else:
                 result_first_half = []
+            now = int(time.time())
+            remaining = timeout - (now - begin_time)
+            if remaining < 0:
+                raise ServiceResponseTimeoutError("Service response time out")
             batch_response_second_half = self._index_documents_actions(
                 actions=actions[pos:],
                 error_map=error_map,
+                timeout=remaining,
                 **kwargs
             )
-            if batch_response_second_half:
+            if len(batch_response_second_half) > 0:
                 result_second_half = cast(List[IndexingResult], batch_response_second_half.results)
             else:
                 result_second_half = []
@@ -315,3 +282,18 @@ class SearchIndexDocumentBatchingClient(HeadersMixin):
         # type: (*Any) -> None
         self.close()
         self._client.__exit__(*args)  # pylint:disable=no-member
+
+    def _retry_action(self, action):
+        # type: (IndexAction) -> None
+        key = action.additional_properties.get(self._index_key)
+        counter = self._retry_counter.get(key)
+        if not counter:
+            # first time that fails
+            self._retry_counter[key] = 1
+            self._index_documents_batch.enqueue_action(action)
+        elif counter < self._RETRY_LIMIT - 1:
+            # not reach retry limit yet
+            self._retry_counter[key] = counter + 1
+            self._index_documents_batch.enqueue_action(action)
+        else:
+            self._fail_callback(action)