+++ sdk/search/azure-search-documents/tests/async_tests/test_batching_client_async.py
@@ -16,16 +16,16 @@ CREDENTIAL = AzureKeyCredential(key="test_api_key")
 
 class TestSearchBatchingClientAsync(object):
     async def test_search_index_document_batching_client_kwargs(self):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=100, batch_size=100)
+        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=100)
 
-        assert client.batch_size == 100
+        assert client.batch_size == 1000
         assert client._window == 100
         assert client._auto_flush
         await client.close()
 
 
     async def test_batch_queue(self):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL)
+        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False)
 
         assert client._index_documents_batch
         await client.add_upload_actions(["upload1"])
@@ -39,49 +39,22 @@ class TestSearchBatchingClientAsync(object):
         assert len(client.actions) == 7
 
 
-    async def test_succeeded_queue(self):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL)
-
-        assert client._index_documents_batch
-        await client.add_upload_actions(["upload1"])
-        await client.add_delete_actions(["delete1", "delete2"])
-        await client.add_merge_actions(["merge1", "merge2", "merge3"])
-        await client.add_merge_or_upload_actions(["merge_or_upload1"])
-        actions = await client._index_documents_batch.dequeue_actions()
-        await client._index_documents_batch.enqueue_succeeded_actions(actions)
-        assert len(client.succeeded_actions) == 7
-
-
-    async def test_failed_queue(self):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL)
-
-        assert client._index_documents_batch
-        await client.add_upload_actions(["upload1"])
-        await client.add_delete_actions(["delete1", "delete2"])
-        await client.add_merge_actions(["merge1", "merge2", "merge3"])
-        await client.add_merge_or_upload_actions(["merge_or_upload1"])
-        actions = await client._index_documents_batch.dequeue_actions()
-        await client._index_documents_batch.enqueue_failed_actions(actions)
-        assert len(client.failed_actions) == 7
-
-
     @mock.patch(
-        "azure.search.documents._internal.aio._search_index_document_batching_client_async.SearchIndexDocumentBatchingClient.flush"
+        "azure.search.documents._internal.aio._search_index_document_batching_client_async.SearchIndexDocumentBatchingClient._process_if_needed"
     )
-    async def test_flush_if_needed(self, mock_flush):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=1000, batch_size=2)
+    async def test_process_if_needed(self, mock_process_if_needed):
+        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=1000, auto_flush=False)
 
         await client.add_upload_actions(["upload1"])
         await client.add_delete_actions(["delete1", "delete2"])
-        assert mock_flush.called
-        await client.close()
+        assert mock_process_if_needed.called
 
 
     @mock.patch(
         "azure.search.documents._internal.aio._search_index_document_batching_client_async.SearchIndexDocumentBatchingClient._cleanup"
     )
     async def test_context_manager(self, mock_cleanup):
-        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL) as client:
+        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
             await client.add_upload_actions(["upload1"])
             await client.add_delete_actions(["delete1", "delete2"])
         assert mock_cleanup.called